backend/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.4.4
Flask-Cors==3.0.10
PyMySQL==1.0.3
python-dotenv==1.0.0
